Moved data_reader base class up in the directory hierarchy.
