+++ include/lbann/data_ingestion/readers/CMakeLists.txt
@@ -26,7 +26,6 @@
 # Add the headers for this directory
 set_full_path(THIS_DIR_HEADERS
   compound_data_reader.hpp
-  data_reader.hpp
   metadata.hpp
   # Data readers
   data_reader_cifar10.hpp
