+++ include/lbann/data_ingestion/readers/data_reader_node2vec.hpp
@@ -27,7 +27,7 @@
 #ifndef LBANN_DATA_READERS_NODE2VEC_HPP_INCLUDED
 #define LBANN_DATA_READERS_NODE2VEC_HPP_INCLUDED
 
-#include "data_reader.hpp"
+#include "lbann/data_ingestion/data_reader.hpp"
 #ifdef LBANN_HAS_LARGESCALE_NODE2VEC
 
 namespace lbann {
