+++ include/lbann/data_ingestion/readers/data_reader_numpy.hpp
@@ -30,7 +30,7 @@
 #ifndef LBANN_DATA_READER_NUMPY_HPP
 #define LBANN_DATA_READER_NUMPY_HPP
 
-#include "data_reader.hpp"
+#include "lbann/data_ingestion/data_reader.hpp"
 #include <cnpy.h>
 
 namespace lbann {
