+++ include/lbann/data_ingestion/readers/data_reader_sample_list.hpp
@@ -28,7 +28,7 @@
 #define LBANN_DATA_READER_SAMPLE_LIST_HPP
 
 // #include "lbann_config.hpp"
-#include "lbann/data_ingestion/readers/data_reader.hpp"
+#include "lbann/data_ingestion/data_reader.hpp"
 #include <conduit/conduit.hpp>
 
 namespace lbann {
