+++ src/data_coordinator/data_coordinator.cpp
@@ -26,8 +26,8 @@
 
 #include <lbann/comm_impl.hpp>
 #include <lbann/data_coordinator/data_coordinator.hpp>
+#include <lbann/data_ingestion/data_reader.hpp>
 #include <lbann/data_ingestion/data_store_conduit.hpp>
-#include <lbann/data_ingestion/readers/data_reader.hpp>
 #include <lbann/execution_algorithms/execution_context.hpp>
 #include <lbann/trainers/trainer.hpp>
 #include <lbann/utils/dim_helpers.hpp>
