+++ src/data_coordinator/unit_test/buffered_data_coordinator_test.cpp
@@ -31,7 +31,7 @@
 
 #include <lbann/base.hpp>
 #include <lbann/data_coordinator/buffered_data_coordinator.hpp>
-#include <lbann/data_ingestion/readers/data_reader.hpp>
+#include <lbann/data_ingestion/data_reader.hpp>
 // #include <lbann/data_ingestion/readers/data_reader_synthetic.hpp>
 #include <lbann/utils/memory.hpp>
 #include <lbann/utils/threads/thread_pool.hpp>
