+++ src/data_ingestion/readers/CMakeLists.txt
@@ -25,7 +25,6 @@
 ################################################################################
 # Add the source files for this directory
 set_full_path(THIS_DIR_SOURCES
-  data_reader.cpp
   metadata.cpp
   data_reader_cifar10.cpp
   data_reader_csv.cpp
