+++ src/utils/lbann_library.cpp
@@ -27,7 +27,7 @@
 #include "lbann/utils/lbann_library.hpp"
 #include "lbann/comm.hpp"
 #include "lbann/comm_impl.hpp"
-#include "lbann/data_ingestion/readers/data_reader.hpp"
+#include "lbann/data_ingestion/data_reader.hpp"
 #include "lbann/models/model.hpp"
 #include "lbann/utils/exception.hpp"
